+++ apps/governance/pages/index.tsx
@@ -19,7 +19,7 @@ import {
   getAllSnackbars,
 } from "evmos-wallet";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { TermOfServices, Container, Consent } from "ui-helpers";
+import { TermOfServices, Container } from "ui-helpers";
 
 function SnackbarsInternal() {
   const valueRedux = useSelector((state: StoreType) => getAllSnackbars(state));
@@ -46,7 +46,6 @@ export default function Home() {
               <HeadComponent />
               <GoogleAnalytics />
               <main>
-                <Consent />
                 <TermOfServices />
                 <InformationBanner
                   dismissible={true}