+++ packages/ui-helpers/src/index.tsx
@@ -29,5 +29,4 @@ export { ViewExplorer } from "./ViewExplorer";
 export { BarContainer } from "./votingBar/BarContainer";
 export { Header } from "./Header";
 export { InformationBanner } from "./InformationBanner";
-export { Consent } from "./termsOfServices/Consent";
-export { useConsent } from "./termsOfServices/useConsent";
+export { ConsentModal } from "./termsOfServices/ConsentModal";